requirements/base.txt
Django==3.2
django-environ==0.4.5
django-simpleui==2021.6.2
django-tinymce==3.3.0
django-braces==1.14.0
django-guardian==2.4.0
django-import-export==2.5.0
pyecharts==1.9.0
zipstream==1.1.4
psycopg2